move trtllm-attention QKV-fusion tests out of standalone-shipped file
Commit 1297aed3be added three tests that depend on the TRT-LLM attention
backend into ``test_gemm_fusion.py``:

  * ``test_fuse_qkv_with_trtllm_cache_insertion``
  * ``test_fuse_qkv_gqa_with_trtllm_cache_insertion``
  * ``test_fuse_qkv_passthrough_with_rope`` (+ ``QKVRopeAttentionModel``)

That file is part of the standalone auto_deploy package, so when the
package is built the tests run with ``TRTLLM_AVAILABLE=False`` and:

  - import ``KvCacheConfig`` directly from ``tensorrt_llm.llmapi.llm_args``
    (collection error: ``ModuleNotFoundError: No module named 'tensorrt_llm'``)
  - call ``torch.ops.auto_deploy.trtllm_attention_mha_with_cache``, which
    is only registered when TRT-LLM is present.

Move the three tests + ``QKVRopeAttentionModel`` into
``test_gemm_fusion_trtllm.py``, which is already in ``EXCLUDE_TEST_FILES``
in ``examples/auto_deploy/llmc/create_standalone_package.py`` for exactly
this reason.  ``QKVAttentionModel`` and ``_get_narrow_nodes`` are imported
from ``test_gemm_fusion`` to avoid duplication.

Drop the unused ``KvCacheConfig`` and ``CachedSequenceInterface`` imports
from ``test_gemm_fusion.py``.

Verified locally:
  - standalone package suite: 5/5 pass (1372 inner tests, 0 errors)
  - in-repo run: ``test_gemm_fusion.py`` + ``test_gemm_fusion_trtllm.py``
    -> 30 passed, 2 skipped, 2 xfailed, 2 xpassed
