+++ tests/unittest/auto_deploy/singlegpu/transformations/library/test_gemm_fusion_trtllm.py
@@ -10,8 +10,7 @@ package excludes this file via ``EXCLUDE_TEST_FILES`` in
 import operator
 
 import torch
-import torch.nn as nn
-from test_gemm_fusion import TestModel  # type: ignore
+from test_gemm_fusion import QKVAttentionModel, _get_narrow_nodes  # type: ignore
 
 import tensorrt_llm._torch.auto_deploy.custom_ops  # noqa: F401 (registers torch_attention op)
 from tensorrt_llm._torch.auto_deploy._compat import KvCacheConfig
@@ -32,75 +31,6 @@ def _count_split_output_nodes(gm):
             if isinstance(source, torch.fx.Node) and source.op == "call_function":
                 count += 1
     return count
-
-
-class QKVAttentionModel(TestModel):
-    """Model with separate Q, K, V projections feeding into torch_attention.
-
-    Mimics the attention pattern in transformer models where Q, K, V are
-    projected from the same input. fuse_gemms_mixed_children should fuse the
-    3 projections into one GEMM with 3 narrow views.
-    """
-
-    def __init__(
-        self,
-        batch_size=2,
-        seq_len=8,
-        hidden_size=64,
-        num_heads=4,
-        num_kv_heads=None,
-    ):
-        super().__init__()
-        if num_kv_heads is None:
-            num_kv_heads = num_heads
-        self.batch_size = batch_size
-        self.seq_len = seq_len
-        self.hidden_size = hidden_size
-        self.num_heads = num_heads
-        self.num_kv_heads = num_kv_heads
-        self.head_dim = hidden_size // num_heads
-
-        self.q_proj = nn.Linear(hidden_size, num_heads * self.head_dim, bias=False)
-        self.k_proj = nn.Linear(hidden_size, num_kv_heads * self.head_dim, bias=False)
-        self.v_proj = nn.Linear(hidden_size, num_kv_heads * self.head_dim, bias=False)
-        self.o_proj = nn.Linear(hidden_size, hidden_size, bias=False)
-
-    def get_input(self, **kwargs):
-        return torch.randn(self.batch_size, self.seq_len, self.hidden_size, **kwargs)
-
-    @property
-    def keys_to_pop(self):
-        return ("q_proj.weight", "k_proj.weight", "v_proj.weight")
-
-    @property
-    def num_gemms_after_fusion(self) -> int:
-        return 2  # 1 fused QKV + 1 o_proj
-
-    @property
-    def expected_narrow_count(self) -> int:
-        return 3  # Q, K, V slices
-
-    def forward(self, x):
-        b, s, _ = x.shape
-        q = self.q_proj(x).view(b, s, self.num_heads, self.head_dim)
-        k = self.k_proj(x).view(b, s, self.num_kv_heads, self.head_dim)
-        v = self.v_proj(x).view(b, s, self.num_kv_heads, self.head_dim)
-
-        attn = torch.ops.auto_deploy.torch_attention.default(
-            q,
-            k,
-            v,
-            attn_mask=None,
-            dropout_p=0.0,
-            is_causal=True,
-            scale=None,
-            sinks=None,
-            sliding_window=None,
-            logit_cap=None,
-            layout="bsnd",
-        )
-        out = attn.reshape(b, s, self.hidden_size)
-        return self.o_proj(out)
 
 
 @torch.inference_mode()
@@ -132,7 +62,7 @@ def test_fuse_qkv_with_trtllm_cache_insertion():
     cm = CachedSequenceInterface(
         max_seq_len=64,
         max_batch_size=4,
-        max_num_tokens=64 * 4,
+        max_num_tokens=256,
         device="cuda",
         kv_cache_config=kv_cache_config,
     )
@@ -160,6 +90,11 @@ def test_fuse_qkv_with_trtllm_cache_insertion():
         f"Expected 1 prepare_metadata node, got {len(prep_meta_nodes)}"
     )
 
+    # The QKV split (3 getitems) must survive cache insertion; cache insertion
+    # itself can add more getitems (e.g. from the metadata-prep tuple), so
+    # require at least 3 rather than exactly 3.
+    assert _count_split_output_nodes(gm) >= 3
+
 
 @torch.inference_mode()
 def test_fuse_qkv_gqa_with_trtllm_cache_insertion():
@@ -184,8 +119,33 @@ def test_fuse_qkv_gqa_with_trtllm_cache_insertion():
 
     assert _count_split_output_nodes(gm) == 3
 
-    # Verify the fused GEMM produces 3 split outputs (Q, K, V)
-    assert _count_split_output_nodes(gm) == 3
+    # Asymmetric Q/K/V split sizes must be preserved.  The fusion may emit
+    # either ``torch.narrow`` ops or a ``split_with_sizes``-style tuple split
+    # depending on the dtype path.  In the split-tuple form the getitem nodes
+    # may lack ``meta['val']`` (the closure isn't shape-propagated), but the
+    # downstream ``view`` they feed into has the right shape — read from there.
+    narrow_sizes = sorted([n.args[3] for n in _get_narrow_nodes(gm)])
+    if not narrow_sizes:
+        getitem_sizes = []
+        for n in gm.graph.nodes:
+            if (
+                n.op == "call_function"
+                and n.target is operator.getitem
+                and isinstance(n.args[0], torch.fx.Node)
+                and n.args[0].op == "call_function"
+            ):
+                val = n.meta.get("val")
+                if val is None:
+                    # Fall back to the view consumer's shape: (B, S, n_heads, head_dim)
+                    for user in n.users:
+                        uval = user.meta.get("val")
+                        if uval is not None and len(uval.shape) >= 4:
+                            getitem_sizes.append(int(uval.shape[-1] * uval.shape[-2]))
+                            break
+                else:
+                    getitem_sizes.append(int(val.shape[-1]))
+        narrow_sizes = sorted(getitem_sizes)
+    assert narrow_sizes == [32, 32, 64], f"Unexpected QKV split sizes for GQA: {narrow_sizes}"
 
     kv_cache_config = KvCacheConfig(
         tokens_per_block=32,
@@ -195,7 +155,7 @@ def test_fuse_qkv_gqa_with_trtllm_cache_insertion():
     cm = CachedSequenceInterface(
         max_seq_len=64,
         max_batch_size=4,
-        max_num_tokens=64 * 4,
+        max_num_tokens=256,
         device="cuda",
         kv_cache_config=kv_cache_config,
     )
@@ -211,3 +171,96 @@ def test_fuse_qkv_gqa_with_trtllm_cache_insertion():
         if is_op(n, torch.ops.auto_deploy.trtllm_attention_mha_with_cache.default)
     ]
     assert len(cached_attn_nodes) == 1
+
+
+class QKVRopeAttentionModel(QKVAttentionModel):
+    """``QKVAttentionModel`` with HF-style explicit cos/sin RoPE on Q and K.
+
+    This shape is what ``fuse_rope_into_trtllm_attention`` looks for: a
+    ``torch_rope_with_explicit_cos_sin`` node sits between Q/K projections
+    and the ``torch_attention`` call, and Q/K/V all trace back to the same
+    fused QKV GEMM output so the QKV-passthrough leg can fire.
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        # ``fuse_rope_into_trtllm_attention`` reads cos/sin as 2D
+        # ``[max_pos, head_dim]`` buffers (HF-style table).
+        self.register_buffer(
+            "rope_cos_table",
+            torch.zeros(self.seq_len, self.head_dim),
+            persistent=False,
+        )
+        self.register_buffer(
+            "rope_sin_table",
+            torch.zeros(self.seq_len, self.head_dim),
+            persistent=False,
+        )
+
+    def forward(self, x):
+        b, s, _ = x.shape
+        # Index cos/sin tables by position so the extractor can trace through
+        # ``aten.index.Tensor`` back to the underlying 2D table.
+        position_ids = torch.arange(s, device=x.device)
+        cos = self.rope_cos_table[position_ids].unsqueeze(0).expand(b, -1, -1)
+        sin = self.rope_sin_table[position_ids].unsqueeze(0).expand(b, -1, -1)
+        q = self.q_proj(x).view(b, s, self.num_heads, self.head_dim)
+        k = self.k_proj(x).view(b, s, self.num_kv_heads, self.head_dim)
+        v = self.v_proj(x).view(b, s, self.num_kv_heads, self.head_dim)
+        q, k = torch.ops.auto_deploy.torch_rope_with_explicit_cos_sin.default(q, k, cos, sin, 2)
+        attn = torch.ops.auto_deploy.torch_attention.default(
+            q,
+            k,
+            v,
+            attn_mask=None,
+            dropout_p=0.0,
+            is_causal=True,
+            scale=None,
+            sinks=None,
+            sliding_window=None,
+            logit_cap=None,
+            layout="bsnd",
+        )
+        out = attn.reshape(b, s, self.hidden_size)
+        return self.o_proj(out)
+
+
+@torch.inference_mode()
+def test_fuse_qkv_passthrough_with_rope():
+    """Exercise the QKV-passthrough leg of fuse_rope_into_trtllm_attention.
+
+    Builds a graph where Q/K/V come from a fused QKV GEMM and Q/K go
+    through ``torch_rope_with_explicit_cos_sin`` before attention.  After
+    rope fusion the ``torch_attention`` node must carry the
+    ``_trtllm_fused_qkv`` marker plus non-zero head/dim hints — proving
+    that the passthrough actually fired.
+    """
+    model = QKVRopeAttentionModel(hidden_size=64, num_heads=4).to(
+        device="cuda", dtype=torch.float16
+    )
+    x = model.get_input(device="cuda", dtype=torch.float16)
+
+    gm = torch_export_to_gm(model, args=(x,), clone=True)
+
+    gm = InferenceOptimizer(
+        None,
+        {
+            "fuse_gemms_mixed_children": {"stage": "post_load_fusion"},
+            "fuse_rope_into_trtllm_attention": {"stage": "post_load_fusion"},
+        },
+    )(None, gm)
+
+    # The torch_attention node should now carry passthrough metadata.
+    torch_attn_nodes = [
+        n for n in gm.graph.nodes if is_op(n, torch.ops.auto_deploy.torch_attention.default)
+    ]
+    assert len(torch_attn_nodes) == 1
+    src = torch_attn_nodes[0]
+    assert src.meta.get("_trtllm_fused_qkv") is True, (
+        "fuse_rope_into_trtllm_attention should have set _trtllm_fused_qkv"
+    )
+    assert src.meta.get("_trtllm_num_heads") == 4
+    assert src.meta.get("_trtllm_num_kv_heads") == 4
+    assert src.meta.get("_trtllm_head_dim") == 16
+    # Q/K/V args must all alias the same fused-QKV node.
+    assert src.args[0] is src.args[1] is src.args[2]
